+++ src/main/java/github/renderbr/hytale/listeners/ChatListener.java
@@ -7,7 +7,9 @@ import com.hypixel.hytale.server.core.Message;
 import com.hypixel.hytale.server.core.event.events.player.PlayerChatEvent;
 import com.hypixel.hytale.server.core.universe.PlayerRef;
 import github.renderbr.hytale.AverageDiscord;
+import github.renderbr.hytale.config.DiscordBridgeConfigurationProvider;
 import github.renderbr.hytale.config.obj.ChannelOutputTypes;
+import github.renderbr.hytale.registries.ProviderRegistry;
 
 public class ChatListener {
 
@@ -26,12 +28,35 @@ public class ChatListener {
 
         if (formattedMessage == null || builtMessage.isEmpty()) {
             // use fallback
-            AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.CHAT, event.getSender().getUsername()
-                    + ": " + event.getContent());
+            var strippedText = stripTextForForbiddenContent(event.getContent());
+            var message = "**" + event.getSender().getUsername() + "**: " + strippedText;
+
+            AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.CHAT, message);
             return;
         }
 
-        AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.CHAT, builtMessage);
+        var strippedText = stripTextForForbiddenContent(builtMessage);
+        AverageDiscord.instance.SendMessageToType(ChannelOutputTypes.CHAT, strippedText);
+    }
+
+    public static String stripTextForForbiddenContent(String text) {
+        if (text == null) return null;
+
+        String newText = text;
+
+        // remove Markdown hyperlinks: [Text](Url)
+        // Matches '[' then any non-bracket chars, then ']', then '(' any non-paren chars, then ')'
+        if (ProviderRegistry.discordBridgeConfigProvider.config.stripLinksInChat) {
+            newText = newText.replaceAll("\\[[^\\]]*\\]\\([^\\)]*\\)", "(stripped url)");
+        }
+
+        // remove Pings: <@...>, <@&...>, <@!...>, @everyone, @here
+        // Matches the @everyone/@here literals OR the <@ID> format
+        if(ProviderRegistry.discordBridgeConfigProvider.config.stripPingsInChat) {
+            newText = newText.replaceAll("@everyone|@here|<@(!|&)?\\d+>", "(stripped ping)");
+        }
+
+        return newText;
     }
 
     public static String recursivelyBuildFormattedMessage(FormattedMessage msg) {